--- a/src/ssec/giirs_ssec2ioda.py
+++ b/src/ssec/giirs_ssec2ioda.py
@@ -18,13 +18,10 @@ import dateutil.parser
 import numpy as np
 import netCDF4 as nc
 
-# Locate src/lib-python/ioda_conv_ncio
-# TODO Replace with setup.py
-IODA_CONV_PATH = "@SCRIPT_LIB_PATH@"
-if Path(IODA_CONV_PATH).is_dir():
-    sys.path.append(IODA_CONV_PATH)
-else:
-    sys.path.append(str((Path(__file__).parent/'..'/'lib-python').resolve()))
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
 
 import ioda_conv_ncio as iconv