--- a/test/testoutput/giirs_ssec_ioda.nc4
+++ b/test/testoutput/giirs_ssec_ioda.nc4
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0ce9d4218035e7697f9763c40c01bdfab1ceef77803240e9b7ef1b8c2da2f361
-size 10302110
+oid sha256:74f992bff9b41910d6878ad934369a26a2862b72982b0d858e18446ac5e9fcb7
+size 10268824