--- a/test/testoutput/giirs_ssec_ioda.nc4
+++ b/test/testoutput/giirs_ssec_ioda.nc4
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9ce5dee45fc27e04e56cb95bd34842a713eca198c958d56175a2a5972a8a0ca6
-size 62719
+oid sha256:99a5ae67feab4126af0260b1d48e2882ad746960b2ff12b0beb8e40f581d716f
+size 62720