This is a performance analysis and optimization for the new `giirs_ssec2ioda.py` converter.  The GIIRS converter reads and collates the observations from a set of several thousand individual scan files for a typical 6Hr window.   

Initial testing showed a single 6-hr window took approximately 60+ mins to collate 1200 obs files and convert to IODA and displayed a non-linear scaling with the number of files processed.  Profiling short runs shows the overwhelming majority of time is spent processing the data read from the files in the `readInFiles` method.  Detailed line profiling identified the majority of the time was spent in the for loops converting radiance units to brightness temperature.

 ![giirs_ssec_orig](https://user-images.githubusercontent.com/14096120/92046334-8945d880-ed3f-11ea-8a9e-2d49c1c56944.png)

These calculations can be vectorized by refactoring the constant terms from the radiance conversion equation and expanding the conversion as an element-wise matrix manipulation rather than as nested for loops.  

After eliminating the radiance conversion overhead, the remaining time can be attributed mainly with the strategy of incrementally re-allocating an `NcVector` for each meta-variable and for each data variable for each input file.  This strategy is quadratic and working with `NcVectors` directly is often 100x slower than numpy arrays.  I modified the strategy with the knowledge that the maximum observations are `128*numfiles`, we can pre-allocate ordinary numpy arrays and matrices for the observations, datetimes, and metadata, etc.  Then any excess space is removed at the end.

Also note:
* Constant vectors `K1` and `K3` are for the radiance conversion and are independent of the obs files and so are moved outside the main loop to avoid re-computation of any terms not necessary.
* The brightness_temperatures for one entire scan file are computed directly into the corresponding slice of the `obs_vals` matrix which has shape `nChans X nTotalLocs`.
* We use a `mask` to mask out the unselected obs.  This controls the total number of obs per file (which is at most 128), and how many columns of `obs_vals` and other metavars are written for each file processed.
* For obs that pass the first `mask`, but fail to have a valid radiance there is a second `valid = (lwRadiance_apo > 0) & (lwRadiance_apo < 300)` mask which controls if the radiance conversion is applied to that loc or not.
* The radiance conversion uses a `log(1+x)` in the original formulation.   This was replaced with the numpy function `np.log1p()` which is faster and more numerically stable especially for small absolute values of `x` which may lead to numerical underflow and loss of precision in the original calculation.
* Radiance computation intermediates are now all at float32 precision.

After applying optimizations, the time for `readInFiles` over a 20 file test set has decreased from `99.0s -> 0.441s` and from 99% of runtime to 20%.  With the remainder of the time spent writing the ioda file in `ioda_conv_ncio.py`.  The resulting file is ~1.2Gb per 6hr period, so approximately 1.5s to write a file this big seems at least reasonable.  I did not explore any further optimization into those routines.

![giirs_ssec_improved](https://user-images.githubusercontent.com/14096120/92047618-c95a8a80-ed42-11ea-90ed-a814a48b9ab9.png)

Overall runtime for a 236 file test set (approx 1hr windows) with 27251 obs:
* **Original**: 370.451s
* **Improved**: 3.499s
* **Speedup**: 106x

Note: after re-arranging the radiance conversions some of the values changed by up to `5.0x10^-5` in relative magnitude, which while small, was enough to cause the tests to fail.   I had to update the reference file with the new values.  I expect these differences are because all the intermediate calculations are now done at 32-bit precisions, and the change to the `np.log1p()` function for the logarithm, which will change precision slightly.  

This is the version that is currently running on S4 and displayed on [nrt.jcsda.org](http://nrt.jcsda.org).

@weihanucar  I would appreciate if you could help review these changes and validate the radiance conversion is still valid and I didn't miss any details.