@markjolah 2020-09-14
Note these are the differences as seen in the output when using the optimized version.  All changes are about at the 9th sig fig, which is related to float32 rounding precision.

```
22: Test command: /home/molah/rapids-home/build/bundle/nrt/iodaconv/_build/bin/iodaconv_comp.sh "netcdf" "/home/molah/rapids-home/build/bundle/nrt/iodaconv/_build/bin/giirs_ssec2ioda.py
             -i testinput/giirs_ssec.nc
             -o testrun/giirs_ssec_ioda.nc4
             -d 2020080100" "giirs_ssec_ioda.nc4"
22: Test timeout computed to be: 1500
22: [GIIRS2IODA] Processing 1/1 #locs:128 -- giirs_ssec.nc
22: [GIIRS2IODA] Processed 128 total locs from 1/1 valid files.
22: DIFFER : VARIABLE : brightness_temperature_115@ObsValue : POSITION : [25] : VALUES : 279.799 <> 279.799
22: DIFFER : VARIABLE : brightness_temperature_118@ObsValue : POSITION : [66] : VALUES : 248.646 <> 248.646
22: DIFFER : VARIABLE : brightness_temperature_363@ObsValue : POSITION : [22] : VALUES : 287.556 <> 287.556
22: DIFFER : VARIABLE : brightness_temperature_373@ObsValue : POSITION : [16] : VALUES : 256.633 <> 256.633
22: DIFFER : VARIABLE : brightness_temperature_397@ObsValue : POSITION : [36] : VALUES : 261.598 <> 261.598
22: DIFFER : VARIABLE : brightness_temperature_467@ObsValue : POSITION : [43] : VALUES : 241.366 <> 241.366
22: DIFFER : VARIABLE : brightness_temperature_566@ObsValue : POSITION : [112] : VALUES : 255.3 <> 255.3
22: DIFFER : VARIABLE : brightness_temperature_656@ObsValue : POSITION : [105] : VALUES : 238.191 <> 238.191
22: DIFFER : VARIABLE : brightness_temperature_662@ObsValue : POSITION : [74] : VALUES : 242.429 <> 242.429
22: Variable                            Group Count          Sum      AbsSum          Min          Max Range         Mean StdDev
22: brightness_temperature_115@ObsValue /         1  3.05176e-05 3.05176e-05  3.05176e-05  3.05176e-05     0  3.05176e-05      0
22: brightness_temperature_118@ObsValue /         1 -3.05176e-05 3.05176e-05 -3.05176e-05 -3.05176e-05     0 -3.05176e-05      0
22: brightness_temperature_363@ObsValue /         1 -3.05176e-05 3.05176e-05 -3.05176e-05 -3.05176e-05     0 -3.05176e-05      0
22: brightness_temperature_373@ObsValue /         1 -3.05176e-05 3.05176e-05 -3.05176e-05 -3.05176e-05     0 -3.05176e-05      0
22: brightness_temperature_397@ObsValue /         1 -3.05176e-05 3.05176e-05 -3.05176e-05 -3.05176e-05     0 -3.05176e-05      0
22: brightness_temperature_467@ObsValue /         1 -1.52588e-05 1.52588e-05 -1.52588e-05 -1.52588e-05     0 -1.52588e-05      0
22: brightness_temperature_566@ObsValue /         1 -3.05176e-05 3.05176e-05 -3.05176e-05 -3.05176e-05     0 -3.05176e-05      0
22: brightness_temperature_656@ObsValue /         1  1.52588e-05 1.52588e-05  1.52588e-05  1.52588e-05     0  1.52588e-05      0
22: brightness_temperature_662@ObsValue /         1 -1.52588e-05 1.52588e-05 -1.52588e-05 -1.52588e-05     0 -1.52588e-05      0
```