@markjolah 2020-09-14
Since this has still not been reviewed, I cleaned up some more items to improve the testing and flexibility, and to fix several bugs I have encountered in production use.  Notable changes:
* Use original `giirs-fy4a-test.nc` file instead of new file I had added.  The original test file has fewer detectors and channels, which makes the test data significantly smaller.  This accomplishes one of my objectives for the testing-code sprint.
* New class name `Giirs2Ioda` is more informative.  Remove references to `longwave_rad` which is not used by the UFO/CRTM operator for GIIRS.
* Allow converter to determine: num detectors, num channels, and wavenumbers from input data files rather than hard-coding them.  This change necessitates moving the construction of the data structures inside the main for loop over data files.  On reading the first successful file, we initialize numpy data structures large enough to hold all remaining data and metadata.  Also, we pre-compute the radiance to brightness temp conversion constants which are identical for each file.
* Check that the assumption that all files in the sequence have the same number of channels and detectors, and throw a RuntimeError if violated.
* Allow for missing metadata names when the input files don't have them. This happens with the test data which is missing a `Cloud_Fraction` variable but this is optional, so we don't want to fail.
* Check for critical missing metadata name: 'IRLW_SatelliteZenith' and drop data files without this variable.  Normally, these are corrupted or empty files.
* Promote the `SAT_ZEN_MAX` filtering constant to a class-level constant.
* Ensure the `scan_position` is an integer and is set correctly for the number of detectors.