@markjolah 2020-09-17
Very interesting.  Fixing the `scan_position@MetaData` to be an `int32` causes the HofX to crash which was the source of the recent failures on the NRT website.  I will revert to a `float32` in the feature/nrt branch until the issue can be resolved.  I will link that issue here when ready.