@markjolah 2020-09-17
See linked issue: [ufo # 1285](https://github.com/JCSDA/ufo/issues/1285)