@markjolah 2020-09-17
@weihanucar If you approve of the PR, you need to click "start a review" and click "approve" otherwise your response without approval causes github to think you've asked a question.