@weihanucar 2020-09-20
There is a bug in the python converter script:
self.xferLocMdataVar_detector(ncd, numFiles, detector_array*1.0, 'scan_position', LocMdata, 'float', mask)
It should be :
self.xferLocMdataVar_detector(ncd, numFiles, detector_array, 'scan_position', LocMdata, 'integer', mask)