@weihanucar 2020-09-20
Hi Mark,

There is a bug in the python converter script:
self.xferLocMdataVar_detector(ncd, numFiles, detector_array*1.0, 'scan_position', LocMdata, 'float', mask)
It should be :
self.xferLocMdataVar_detector(ncd, numFiles, detector_array, 'scan_position', LocMdata, 'integer', mask)




Wei
 
From: Mark J Olah
Date: 2020-09-17 09:21
To: JCSDA/ioda-converters
CC: weihanucar; Mention
Subject: Re: [JCSDA/ioda-converters] Feature giirs ssec performance (#322)
Very interesting. Fixing the scan_position@MetaData to be an int32 causes the HofX to crash which was the source of the recent failures on the NRT website. I will revert to a float32 in the feature/nrt branch until the issue can be resolved. I will link that issue here when ready.
—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub, or unsubscribe.