@markjolah 2020-09-21
@weihanucar Yes, as you saw we discovered that issue after I updated the PR based on reviewers' comments.  The UFO/CRTM interface somehow requires `scan_postion@MetaData` to be `float32` or it cannot be found.  I've reverted back to using a float32 for scan_position, which I think should solve the problem, and is working for the NRT website now.