@srherbener 2020-09-22
The insistence of the `scan_position@MetaData` to be a float is a design flaw in the current ObsSpace implementation. The container for the data (ObsSpaceContainer) is templated and if you call get_db() to read data from the ObsSpace using the float signature (which is probably what is being done in UFO), it looks in the float version of ObsSpaceContainer. This works if you make `scan_position@MetaData` in the ioda file a float (so that is get loaded into the float version of ObsSpaceContainer).

Changing `scan_position@MetaData` to an int in the file caused it to get loaded into the int version of ObsSapceContainer of which the get_db() with the float signature cannot find. It would also work to change both the ioda file and the UFO get_db() call (to the int signature), but I think we should wait until the new re-factored ObsSpace is in place before doing this. The new ObsSpace will give the caller the ability to query the type of the data and then call the appropriate templated read function.